add-on divided in files to improve maintenance
